pyproject.toml
[project]
name = "event-management-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2.5",
    "django-allauth[socialaccount]>=65.11.0",
    "django-browser-reload>=1.21.0",
    "django-countries>=7.6.1",
    "django-debug-toolbar>=6.0.0",
    "django-environ>=0.12.0",
    "django-htmx>=1.23.2",
    "honcho>=2.0.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "qrcode>=8.2",
    "requests>=2.32.5",
]
